requirements.txt
streamlit
requests
pandas
pycryptodome